Move non remote fetcher to dedicated file
