+++ src/services/index.ts
@@ -4,14 +4,18 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-export * as ExploreSvc from './explore';
-export type * from './explore';
+
+export * as AppLocalSvc from './app-local';
+export type * from './app-local';
 
 export * as AppsMetadataSvc from './apps-metadata';
 export type * from './apps-metadata';
 
 export * as DirectorySvc from './directory';
 export type * from './directory';
+
+export * as ExploreSvc from './explore';
+export type * from './explore';
 
 export * as StudySvc from './study';
 export type * from './study';
